+++ packages/moengage_flutter/moengage_flutter_web/lib/constants.dart
@@ -15,6 +15,8 @@ const String methodSetAliasSDK = 'update_unique_user_id';
 const String methodLogoutSDK = 'destroy_session';
 const String methodEnableSDK = 'enableSdk';
 const String methodDisableSDK = 'disableSdk';
+const String methodEnableDataTracking = 'enableDataTracking';
+const String methodDisableDataTracking = 'disableDataTracking';
 
 // Keys Constants
 String keyEventName = 'eventName';